Create CMake project and basic Hello World program
I decided to use CMake as the build engine for this project when I
learned that it was capable of generating Visual Studio solution (*.sln)
and project (*.vcxproj) files. I'm all for cross-platform compatibility,
and since I already know that CMake works on Linux I'm going to try to
use it for development on Windows as well.

I created a very basic CMakeLists file and a simple "Hello World" C++
program, then configured CMake to build the Hello World program. Upon
running CMake, a number of Visual Studio project and solution files are
generated, in addition to a number of CMake meta files. Obviously we
don't want to commit these, so I added a beefy .gitignore as well.

The current CMake configuration is, as I said, very basic. I'm no CMake
guru (at least not yet), but eventually I'd like to get to the point
where CMake doesn't clutter the project root directory with all of those
project and meta files. Instead, I'd prefer that the project and
solution files be placed in a directory like `proj/` and the CMake
temporary files be placed in `tmp/`.