Fix CMake placing project files in root directory
Last commit I was bemoaning the fact that CMake was cluttering up the
project root directory with tons of Visual Studio project files (*.sln
and *.vcxproj) and temporary files (CMakeCache.txt and CMakeFiles/). I
was looking for a way to redirect this output to a subdirectory so that
the root directory of the project stays tidy, and I finally found one.

It's worth pointing out that there are a number of red herrings on the
Internet for the Google searches related to the problem I was trying to
solve, most notably the fact that you can call `set(CMAKE_FOO_DIRECTORY)`
on various CMake pre-programmed directories to redirect the output.
Perhaps you can, but it never worked for me.

I finally learned from the CMakeWiki (link below) that if CMake detects a
CMakeCache.txt file in the working directory (the directory it is told
to run on), it automatically assumes that that is the build directory,
regardless of what the CMakeLists.txt file says. This means that if you
are in `build/` and you run `cmake ..` and the parent directory has a
CMakeCache.txt file, CMake will assume that the build directory is the
parent directory and will place all built files there.

So the resolution here was twofold:
 1) Run CMake from the directory in which you want the build files to be
 placed (I recommend `build/`, though this directory doesn't get committed
 so it doesn't really matter)

 2) Make sure there isn't a CMakeCache.txt file in a directory other
 than the directory you are running CMake from!

 Relevant links:
  - Out-of-source build trees on the CMake Wiki:
    https://cmake.org/Wiki/CMake_FAQ#Out-of-source_build_trees